Added README

+++ src/gateway.py
@@ -1,5 +1,7 @@
 #Derived from https://wiki.python.org/moin/BaseHttpServer#CA-86c92a7a96bb671a2cb7471059da597b6df2666f_1
 
+from __future__ import print_function
+
 import time
 import BaseHTTPServer
 import CGIHTTPServer
@@ -9,6 +11,7 @@ import os
 import urllib2
 import thread
 import socket
+import sys
 
 HOST_NAME = '0:0:0:0:0:0:0:1' # Change this!
 PORT_NUMBER = 8080 # Change this!
@@ -27,11 +30,12 @@ class MyHandler(CGIHTTPServer.CGIHTTPRequestHandler):
         filename = "queue/" + str(int(time.time() * 1000000)) + "_" + s.client_address[0]
         if not os.path.exists(os.path.dirname(filename)):
             os.makedirs(os.path.dirname(filename))
-        fh = open(filename, 'w')
         data = s.rfile.read(int(s.headers.getheader('content-length')))
+        fh = open(filename, 'w')
         fh.write(data.decode())
         s.send_response(204)
         s.end_headers()
+
     def do_PUT(s):
 	do_POST(s)
 
@@ -47,11 +51,15 @@ def processQueue():
                 request = urllib2.Request(NEXT_SERVER, data=data)
                 request.add_header('Content-Type', 'raw')
                 url = opener.open(request)
-                os.remove(filename)
+                print("Status = %s" % url.getcode())
+                if int(url.getcode()/100) == 2:
+                    os.remove(filename)
+                else:
+                    print("Status code from next server was not success", file=sys.stderr)
             else:
                 time.sleep(1)
         except Exception, e:
-            print e
+            print(e, file=sys.stderr)
             time.sleep(1)
 
 class HTTPServerV6(BaseHTTPServer.HTTPServer):
@@ -60,10 +68,10 @@ class HTTPServerV6(BaseHTTPServer.HTTPServer):
 if __name__ == '__main__':
     thread.start_new_thread(processQueue, ())
     httpd = HTTPServerV6((HOST_NAME, PORT_NUMBER), MyHandler)
-    print time.asctime(), "Server Starts - %s:%s" % (HOST_NAME, PORT_NUMBER)
+    print(time.asctime(), "Server Starts - %s:%s" % (HOST_NAME, PORT_NUMBER))
     try:
         httpd.serve_forever()
     except KeyboardInterrupt:
         pass
     httpd.server_close()
-    print time.asctime(), "Server Stops - %s:%s" % (HOST_NAME, PORT_NUMBER)
+    print(time.asctime(), "Server Stops - %s:%s" % (HOST_NAME, PORT_NUMBER))
